cast computation to match format string

+++ smtpd/mta_session.c
@@ -750,7 +750,7 @@ mta_enter_state(struct mta_session *s, int newstate)
 			}
 
 			log_debug("mta: debug: last connection: hanging on for %is",
-			    s->relay->limits->sessdelay_keepalive - s->hangon);
+			    (int)(s->relay->limits->sessdelay_keepalive - s->hangon));
 			s->flags |= MTA_HANGON;
 			runq_schedule(hangon, time(NULL) + 1, NULL, s);
 			break;
